refactor: TimeSlot, BundleInfo 관련 타입 정의 위치 변경 (#211)
