+++ src/shared/types/common.ts
@@ -1,1 +1,1 @@
-// common.ts
+export type TimeSlot = 'DAWN' | 'MORNING' | 'AFTERNOON' | 'EVENING'
